# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimalView API - animal movement tracks and nearby street-level imagery
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Species {
  id: string;
  commonName: string;
  scientificName: string;
  habitat?: string;
}

export interface Study {
  id: string;
  name: string;
  principalInvestigator?: string;
  location?: string;
}

export interface Individual {
  id: string;
  name: string;
  sex?: string;
  nickname?: string;
}

export interface TrackPoint {
  lat: number;
  lon: number;
  timestamp: string;
}

export interface Track {
  studyId: string;
  individualId: string;
  /** real | demo */
  mode: string;
  points: TrackPoint[];
}

export interface MatchImageryRequest {
  /** Search radius in meters */
  radius: number;
  providers: string[];
  points: TrackPoint[];
}

export interface ImageryMatch {
  trackPointIndex: number;
  /** google | mapillary | demo */
  provider: string;
  distanceM: number;
  panoId?: string;
  imageId?: string;
  imageLat?: number;
  imageLon?: number;
  imageDate?: string;
  heading?: number;
  /** high | medium | low */
  confidence: string;
  previewUrl?: string;
}

export type ListSpecies200 = {
  species: Species[];
};

export type ListStudiesParams = {
species: string;
};

export type ListStudies200 = {
  studies: Study[];
};

export type ListIndividualsParams = {
studyId: string;
};

export type ListIndividuals200 = {
  individuals: Individual[];
};

export type GetTrackParams = {
studyId: string;
individualId: string;
from?: string;
to?: string;
};

export type MatchImagery200 = {
  /** real | demo */
  mode: string;
  matches: ImageryMatch[];
};

export type GetProviders200 = {
  google: boolean;
  mapillary: boolean;
  movebank: boolean;
  demoMode: boolean;
};