# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AnimalView API - animal movement tracks and nearby street-level imagery
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /studies:
    get:
      operationId: listStudies
      tags: [catalog]
      summary: List studies for a species
      parameters:
        - in: query
          name: species
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Studies list
          content:
            application/json:
              schema:
                type: object
                required: [studies]
                properties:
                  studies:
                    type: array
                    items:
                      $ref: "#/components/schemas/Study"
  /individuals:
    get:
      operationId: listIndividuals
      tags: [catalog]
      summary: List individuals within a study
      parameters:
        - in: query
          name: studyId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Individuals list
          content:
            application/json:
              schema:
                type: object
                required: [individuals]
                properties:
                  individuals:
                    type: array
                    items:
                      $ref: "#/components/schemas/Individual"
...
components:
  schemas:
...
    Study:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        principalInvestigator:
          type: string
        location:
          type: string
    Individual:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        sex:
          type: string
        nickname:
          type: string
...