# lib/api-zod/src/generated/types/individual.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimalView API - animal movement tracks and nearby street-level imagery
 * OpenAPI spec version: 0.1.0
 */

export interface Individual {
  id: string;
  name: string;
  sex?: string;
  nickname?: string;
}