# lib/api-zod/src/generated/types/study.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimalView API - animal movement tracks and nearby street-level imagery
 * OpenAPI spec version: 0.1.0
 */

export interface Study {
  id: string;
  name: string;
  principalInvestigator?: string;
  location?: string;
}